requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
assemblyai>=0.33.0
httpx>=0.27.0
httpcore>=1.0.0
supabase
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
groq
itsdangerous
